Refactor referral columns to own file
