+++ src/modules/ce/components/client/ClientReferralsTable.tsx
@@ -2,37 +2,21 @@ import { Paper } from '@mui/material';
 import React from 'react';
 import { ColumnDef } from '@/components/elements/table/types';
 import useSafeParams from '@/hooks/useSafeParams';
-import { REFERRAL_COLUMNS } from '@/modules/ce/components/project/ProjectReferralsTable';
+
+import {
+  REFERRAL_COLUMNS,
+  REFERRAL_WITH_PROJECT_COLUMNS,
+} from '@/modules/ce/referralColumns';
 import { getReferralLink } from '@/modules/ce/util';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
-import ProjectTypeChip from '@/modules/hmis/components/ProjectTypeChip';
 import { useFilters } from '@/modules/hmis/filterUtil';
 import {
   CeReferralStatus,
-  CeReferralWithProjectFieldsFragment,
   ClientCeReferralTableFieldsFragment,
   GetClientCeReferralsDocument,
   GetClientCeReferralsQuery,
   GetClientCeReferralsQueryVariables,
 } from '@/types/gqlTypes';
-
-export const REFERRAL_WITH_PROJECT_COLUMNS: {
-  [key: string]: ColumnDef<CeReferralWithProjectFieldsFragment>;
-} = {
-  projectName: {
-    header: 'Project Name',
-    key: 'projectName',
-    render: (referral: CeReferralWithProjectFieldsFragment) =>
-      referral.targetProjectName,
-  },
-  projectType: {
-    header: 'Project Type',
-    key: 'projectType',
-    render: (referral: CeReferralWithProjectFieldsFragment) => (
-      <ProjectTypeChip projectType={referral.targetProjectType} />
-    ),
-  },
-};
 
 const COLUMNS: ColumnDef<ClientCeReferralTableFieldsFragment>[] = [
   { ...REFERRAL_WITH_PROJECT_COLUMNS.projectName, sticky: 'left' },
